requirements.txt
aiofiles==0.7.0
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.11.0
dnspython==2.1.0
easydict==1.9
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
h2==4.1.0
h5py>=3.1.0
hpack==4.0.0
httptools==0.2.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.19.5
opencv-python==4.5.2.54
orjson==3.6.4
pandas==1.2.5
Pillow==8.2.0
priority==2.0.0
prometheus-fastapi-instrumentator==5.7.1
protobuf==3.19.1
pydantic==1.8.2
pyparsing==3.0.4
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tabulate==0.8.9
tensorboardX==2.4
threadpoolctl==3.0.0
toml==0.10.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchvision==0.11.1+cu113
tqdm==4.60.0
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
wsproto==1.0.0
